New way to uniquely identify each subject. Added teacher_name table.

+++ app/src/main/java/com/subreax/schedule/data/local/entitiy/LocalExpandedSubject.kt
@@ -1,12 +1,12 @@
 package com.subreax.schedule.data.local.entitiy
 
 data class LocalExpandedSubject(
-    val id: Int,
+    val id: Long,
     val typeId: String,
     val name: String,
     val place: String,
     val teacher: String,
-    val beginTimeMins: Long,
-    val endTimeMins: Long,
+    val beginTimeMins: Int,
+    val endTimeMins: Int,
     val note: String
 )
