+++ app/src/main/java/com/subreax/schedule/data/repository/schedule/ScheduleRepository.kt
@@ -8,5 +8,5 @@ import com.subreax.schedule.utils.Resource
 interface ScheduleRepository {
     suspend fun getSchedule(owner: ScheduleOwner): Resource<Schedule>
     suspend fun deleteSchedule(owner: ScheduleOwner): Resource<Unit>
-    suspend fun findSubjectById(id: Int): Subject?
+    suspend fun findSubjectById(id: Long): Subject?
 }
